Update Playbtn
Play button updated to specify "run/252490//", the game id for Rust, and server ip port changed to 28119. The previous method of calling the query port 28116 was not returning the correct rust game id and it was failing to launch. Unsure if this was caused by recent rust update or changes to steam.